Working towards a valid getIndex().
The algorithm for determining the index will be to call hashCode() on the
key, take the absolute value of that, convert it to hex, and use the least
significant digit which will be between 0 and F.

With a test key of "testKey", the hashCode() value is -1422475155, the absolute
hex value is 54C93F93, and so the index returned should be 3.

The test fails since getIndex() is not fully implemented.